// src/app/features/dashboard/contact-us/model/contact-us.ts
export interface IContactUs {
  data: IDataContactUs;
}

export interface IDataContactUs {
  id: number;
  en_address: string;
  ar_address: string;
  google_map_link: string;
  en_working_hours: string;
  ar_working_hours: string;
  main_phone: string;
  whatsapp_phone: string;
  main_email: string;
  facebook_link: string;
  instagram_link: string;
  linkedin_link: string;
  twitter_link: string;
  youtube_link: string;
  snapchat_link: string;
  telegram_link: string;
  tiktok_link: string;
  en_footer_text: string;
  ar_footer_text: string;
}